Update load images templates

+++ WhoIsDemo/model/ParamsFaceProcessing.cs
@@ -6,8 +6,8 @@
         public int FACEDET_CONFIDENCE_THRESHOLD { get; set; }
         public int FACEDET_SPEED_ACCURACY_MODE { get; set; }
         public int FACETMPLEXT_SPEED_ACCURACY_MODE { get; set; }
-        public int TRACK_MAX_FACE_SIZE { get; set; }
-        public int TRACK_MIN_FACE_SIZE { get; set; }
+        public float TRACK_MAX_FACE_SIZE { get; set; }
+        public float TRACK_MIN_FACE_SIZE { get; set; }
         public int QUALITY_MODEL { get; set; }
         public int FACE_MAX_DETECT { get; set; }
     }
